Split out isThemeGutenbergFirst
