+++ client/state/themes/selectors/index.js
@@ -1,7 +1,7 @@
 /**
  * External dependencies
  */
-import { includes, intersection, get } from 'lodash';
+import { includes, get } from 'lodash';
 
 /**
  * Internal dependencies
@@ -50,25 +50,7 @@ export { isDownloadableFromWpcom } from 'state/themes/selectors/is-downloadable-
 export { shouldFilterWpcomThemes } from 'state/themes/selectors/should-filter-wpcom-themes';
 export { getJetpackUpgradeUrlIfPremiumTheme } from 'state/themes/selectors/get-jetpack-upgrade-url-if-premium-theme';
 export { getPremiumThemePrice } from 'state/themes/selectors/get-premium-theme-price';
-
-/**
- * Checks if a theme should be customized primarily with Gutenberg.
- *
- * Examples include Template First Themes, which can be determined by the feature
- * global-styles or auto-loading-homepage.
- *
- * @param {object} state   Global state tree
- * @param {string} themeId An identifier for the theme - like
- *                         `independent-publisher-2` or `maywood`.
- * @returns {boolean} True if the theme should be edited with gutenberg.
- */
-export function isThemeGutenbergFirst( state, themeId ) {
-	const theme = getTheme( state, 'wpcom', themeId );
-	const themeFeatures = getThemeTaxonomySlugs( theme, 'theme_feature' );
-	const neededFeatures = [ 'global-styles', 'auto-loading-homepage' ];
-	// The theme should have a positive number of matching features to qualify.
-	return !! intersection( themeFeatures, neededFeatures ).length;
-}
+export { isThemeGutenbergFirst } from 'state/themes/selectors/is-theme-gutenberg-first';
 
 const emptyList = [];
